Refactor app structure and implement multi-page navigation
- Migrate existing app.py to a multi-page Streamlit application
- Create separate topic pages for different analysis areas
- Add home.py as the default landing page
- Implement navigation using st.navigation()
- Move previous single-file analysis code to archive/price_analysis.py
- Simplify page configuration and improve overall app structure
