+++ home.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+st.title("Home")
+st.write("This is the home page of the NEM Dashboard.")
