+++ topics/Topic-3-Renewable-Integration.py
@@ -1,0 +1,10 @@
+import os
+import streamlit as st
+import pandas as pd
+import altair as alt
+import warnings
+import datetime
+warnings.filterwarnings("ignore", message="Could not infer format, so each element will be parsed individually")
+
+
+st.title("Topic 3: Renewable Integration")
